+++ src/aoc_cxx/day02.cpp
@@ -10,9 +10,9 @@ namespace aoc {
 
 namespace fs = std::filesystem;
 
-RegisterCommand day02("day02", {{"day02.txt", 1, 2}}, [](fs::path filename) -> std::tuple<size_t, size_t> {
-  size_t part1;
-  size_t part2;
+RegisterCommand day02("day02", {{"day02.txt", 1, 2}}, [](fs::path filename) -> std::tuple<uint64_t, uint64_t> {
+  uint64_t part1;
+  uint64_t part2;
   rust::day02(filename.string(), part1, part2);
   spdlog::debug("result from rust: part1 {} part2 {}", part1, part2);
   return {part1, part2};
